refactor: move WorldBlockersController to its own assembly
- Interfaces AND implementation were moved to their own separate assemblies.
- Implementation side, had to remove the Environment references and replace them with dependency injection
- No extra logic changes had to be performed
- Fixed all assembly definition issues that arised from the refactor
