+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Font/Tests/FontTests.asmdef
@@ -16,7 +16,8 @@
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
     "GUID:97d8897529779cb49bebd400c7f402a6",
     "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
-    "GUID:5289a39f46259954ca74010d7c9ded16"
+    "GUID:5289a39f46259954ca74010d7c9ded16",
+    "GUID:edd6917fa0e214e44b6592eeab419ce7"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
