+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/Tests/UUIDComponentTests.asmdef
@@ -5,16 +5,16 @@
         "GUID:e555144732b726c4cba5b0f6337fea75",
         "GUID:54b09a4a10c111e46a817873fcea53fe",
         "GUID:760a1d365aad58044916992b072cf2a6",
-        "GUID:4973650d2444c4561a15d50f24d91cd9",
-        "GUID:b1087c5731ff68448a0a9c625bb7e52d",
-        "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
-        "GUID:4307f53044263cf4b835bd812fc161a4",
-        "GUID:1a483cdd86db6e641ad77c7046dd035f",
-        "GUID:6055be8ebefd69e48b49212b09b47b2f",
-        "GUID:27619889b8ba8c24980f49ee34dbb44a",
-        "GUID:0acc523941302664db1f4e527237feb3",
-        "GUID:a881b57670b1d2747a6d7f9e32b63230",
-        "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
+      "GUID:4973650d2444c4561a15d50f24d91cd9",
+      "GUID:b1087c5731ff68448a0a9c625bb7e52d",
+      "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
+      "GUID:4307f53044263cf4b835bd812fc161a4",
+      "GUID:1a483cdd86db6e641ad77c7046dd035f",
+      "GUID:6055be8ebefd69e48b49212b09b47b2f",
+      "GUID:27619889b8ba8c24980f49ee34dbb44a",
+      "GUID:0acc523941302664db1f4e527237feb3",
+      "GUID:a881b57670b1d2747a6d7f9e32b63230",
+      "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
       "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
       "GUID:1dd0780aa6be12b428c8005d0bee46b8",
       "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
@@ -26,7 +26,8 @@
       "GUID:b308ae946900fa04080222d7daf47d9a",
       "GUID:97d8897529779cb49bebd400c7f402a6",
       "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
-      "GUID:5289a39f46259954ca74010d7c9ded16"
+      "GUID:5289a39f46259954ca74010d7c9ded16",
+      "GUID:edd6917fa0e214e44b6592eeab419ce7"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
