+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6bfaabdcee7d8304d96b0f573482700b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
