+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/Interfaces/IWorldBlockersController.cs
@@ -1,0 +1,13 @@
+﻿using DCL.Rendering;
+
+namespace DCL.Controllers
+{
+    public interface IWorldBlockersController
+    {
+        void Initialize(ISceneHandler sceneHandler, IBlockerInstanceHandler blockerInstanceHandler);
+        void InitializeWithDefaultDependencies(ISceneHandler sceneHandler, ICullingController cullingController);
+        void SetupWorldBlockers();
+        void SetEnabled(bool targetValue);
+        void Dispose();
+    }
+}
