+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/Interfaces/WorldBlockersControllerInterfaces.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: edd6917fa0e214e44b6592eeab419ce7
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
