+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/Tests/BlockerControllerTests.asmdef
@@ -1,5 +1,6 @@
 {
   "name": "BlockerControllerTests",
+  "rootNamespace": "",
   "references": [
     "GUID:cc6bfabbfe87b7949aa248893f89ce37",
     "GUID:9fad297e585ef204c9c2c497d77099f0",
@@ -13,7 +14,10 @@
     "GUID:e555144732b726c4cba5b0f6337fea75",
     "GUID:3f13bd1a926f03a4cab988f06a80c1fd",
     "GUID:97d8897529779cb49bebd400c7f402a6",
-    "GUID:f71b8c619c7cf404ebe3a1fa4c7742fd"
+    "GUID:f71b8c619c7cf404ebe3a1fa4c7742fd",
+    "GUID:554e630085c589a4185494d56f633366",
+    "GUID:edd6917fa0e214e44b6592eeab419ce7",
+    "GUID:b97702e026e0dbd4397dc4013ef9fcb3"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
