+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/Tests/BlockersControllerShould.cs
@@ -32,7 +32,7 @@ public class BlockersControllerShould
         animationHandler.FadeOut(Arg.Any<GameObject>(), Arg.Invoke());
 
         var newBlockerInstanceHandler = new BlockerInstanceHandler();
-        newBlockerInstanceHandler.Initialize(animationHandler);
+        newBlockerInstanceHandler.Initialize(animationHandler, null);
 
         blockerInstanceHandler = newBlockerInstanceHandler;
         blockersParent = new GameObject();
