+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BlockerController/WorldBlockersController.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 554e630085c589a4185494d56f633366
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
