+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/Tests/BuilderInWorldTest.asmdef
@@ -44,7 +44,8 @@
     "GUID:6e6a87f97192ec947993f360104e75b7",
     "GUID:97d8897529779cb49bebd400c7f402a6",
     "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
-    "GUID:5289a39f46259954ca74010d7c9ded16"
+    "GUID:5289a39f46259954ca74010d7c9ded16",
+    "GUID:edd6917fa0e214e44b6592eeab419ce7"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
