+++ unity-renderer/Assets/Scripts/MainScripts/MainScripts.asmdef
@@ -62,7 +62,9 @@
         "GUID:6e6a87f97192ec947993f360104e75b7",
         "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
         "GUID:5289a39f46259954ca74010d7c9ded16",
-        "GUID:f71b8c619c7cf404ebe3a1fa4c7742fd"
+        "GUID:f71b8c619c7cf404ebe3a1fa4c7742fd",
+        "GUID:edd6917fa0e214e44b6592eeab419ce7",
+        "GUID:554e630085c589a4185494d56f633366"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
